threw away lots of old code, imlemented collection layout to display tiles

+++ PHOTZLE/Source/Categories/UIImage+Resize.m
@@ -78,7 +78,7 @@
             break;
             
         default:
-            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %d", contentMode];
+            [NSException raise:NSInvalidArgumentException format:@"Unsupported content mode: %ld", (long)contentMode];
     }
     
     CGSize size = CGSizeMake(self.size.width * ratio, self.size.height * ratio);
